refactor: move bin to src
