+++ .gitignore
@@ -9,3 +9,4 @@ docs
 .idea
 /docker-compose.override.yml
 /pnpm-debug.log
+bin
