+++ package.json
@@ -3,7 +3,7 @@
   "version": "3.0.0",
   "description": "Aeternity command line interface",
   "bin": {
-    "aecli": "./bin/aecli.js"
+    "aecli": "src/aecli.js"
   },
   "dependencies": {
     "@aeternity/aepp-sdk": "github:aeternity/aepp-sdk-js#feature/poll-intervals",
@@ -28,7 +28,7 @@
     "url": "https://github.com/aeternity/aepp-cli-js"
   },
   "scripts": {
-    "docs": "scripts/js-to-md.js bin/*.js bin/commands/*.js bin/actions/*.js bin/utils/*.js",
+    "docs": "scripts/js-to-md.js src/*.js src/commands/*.js src/actions/*.js src/utils/*.js",
     "lint": "commitlint --from 23f6888 && standard",
     "test": "mocha --recursive",
     "test:watch": "mocha --recursive --watch",
