+++ src/actions/inspect.js
