+++ src/aecli-inspect.js
