+++ src/commands/chain.js
