+++ src/commands/inspect.js
