+++ test/cli/oracle.test.js
@@ -18,7 +18,7 @@
 import { Crypto } from '@aeternity/aepp-sdk'
 import { before, describe, it } from 'mocha'
 import { configure, executeProgram, plan, ready, WALLET_NAME } from './index'
-import oracleProgramFactory from '../../bin/commands/oracle'
+import oracleProgramFactory from '../../src/commands/oracle'
 
 const executeOracle = args => executeProgram(oracleProgramFactory, args, { withNetworkId: true })
 plan(10000000000000)
